+++ ReactAndroid/src/main/java/com/facebook/react/v8executor/Android.mk
@@ -11,7 +11,7 @@ LOCAL_MODULE := v8executor
 
 LOCAL_SRC_FILES := $(wildcard $(LOCAL_PATH)/*.cpp)
 
-LOCAL_C_INCLUDES := $(LOCAL_PATH) $(THIRD_PARTY_NDK_SRC_DIR)/..
+LOCAL_C_INCLUDES := $(LOCAL_PATH)
 
 LOCAL_CFLAGS += -fvisibility=hidden -fexceptions -frtti
