+++ ReactAndroid/src/main/java/com/facebook/react/v8executor/OnLoad.cpp
@@ -3,14 +3,48 @@
 // This source code is licensed under the MIT license found in the
  // LICENSE file in the root directory of this source tree.
 
+#include <fb/fbjni.h>
+#include <folly/Memory.h>
+#include <jsi/V8Runtime.h>
+#include <jsireact/JSIExecutor.h>
 #include <react/jni/JavaScriptExecutorHolder.h>
 #include <react/jni/JReactMarker.h>
+#include <react/jni/JSLogging.h>
 #include <react/jni/ReadableNativeMap.h>
-
-#include "V8ExecutorFactory.h"
 
 namespace facebook {
 namespace react {
+
+namespace {
+
+class V8ExecutorFactory : public JSExecutorFactory {
+public:
+  V8ExecutorFactory(folly::dynamic&& v8Config) :
+    m_v8Config(std::move(v8Config)) {
+  }
+
+  std::unique_ptr<JSExecutor> createJSExecutor(
+      std::shared_ptr<ExecutorDelegate> delegate,
+      std::shared_ptr<MessageQueueThread> jsQueue) override {
+
+    auto logger = std::make_shared<JSIExecutor::Logger>([](const std::string& message, unsigned int logLevel) {
+                    reactAndroidLoggingHook(message, logLevel);
+    });
+
+    return folly::make_unique<JSIExecutor>(
+      facebook::v8runtime::makeV8Runtime(m_v8Config, logger),
+      delegate,
+      *logger,
+      JSIExecutor::defaultTimeoutInvoker,
+      nullptr);
+  }
+
+private:
+  folly::dynamic m_v8Config;
+};
+
+}
+
 
 // This is not like JSCJavaScriptExecutor, which calls JSC directly.  This uses
 // JSIExecutor with V8Runtime.
@@ -26,7 +60,7 @@ class V8ExecutorHolder
     // Android.
     JReactMarker::setLogPerfMarkerIfNeeded();
     // TODO mhorowitz T28461666 fill in some missing nice to have glue
-    return makeCxxInstance(folly::make_unique<jsi::V8ExecutorFactory>(v8Config->consume()));
+    return makeCxxInstance(folly::make_unique<V8ExecutorFactory>(v8Config->consume()));
   }
 
   static void registerNatives() {
